Tests/LibWeb/Text/expected/HTML/autoplay-policy-blocks-becoming-audible.txt

muted autoplay started: true
paused after unmuting without a gesture: true
scripted audible play() rejected with: NotAllowedError
